Refactor products_data file - move headers related functions to separate file
