+++ saleor/csv/utils/products_data.py
@@ -1,216 +1,17 @@
 import os
-from collections import ChainMap, defaultdict
+from collections import defaultdict
 from typing import TYPE_CHECKING, Dict, List, Optional, Set, Tuple, Union
 
 from django.conf import settings
 from django.db.models import Case, CharField, Value as V, When
 from django.db.models.functions import Concat
 
-from ...channel.models import Channel
 from ...core.utils import build_absolute_uri
-from ...product.models import Attribute
-from ...warehouse.models import Warehouse
+from . import ProductExportFields
 
 if TYPE_CHECKING:
     # flake8: noqa
     from django.db.models import QuerySet
-
-
-class ProductExportFields:
-    """Data structure with fields for product export."""
-
-    HEADERS_TO_FIELDS_MAPPING = {
-        "fields": {
-            "id": "id",
-            "name": "name",
-            "description": "description",
-            "category": "category__slug",
-            "product type": "product_type__name",
-            "charge taxes": "charge_taxes",
-            "product weight": "product_weight",
-            "variant sku": "variants__sku",
-            "variant weight": "variant_weight",
-            "cost price": "variants__cost_price_amount",
-            "variant currency": "variants__currency",
-        },
-        "product_many_to_many": {
-            "collections": "collections__slug",
-            "product images": "images__image",
-        },
-        "variant_many_to_many": {"variant images": "variants__images__image",},
-    }
-
-    PRODUCT_ATTRIBUTE_FIELDS = {
-        "value": "attributes__values__slug",
-        "slug": "attributes__assignment__attribute__slug",
-        "attribute_pk": "attributes__assignment__attribute__pk",
-    }
-
-    PRODUCT_CHANNEL_LISTING_FIELDS = {
-        "channel_pk": "channel_listing__channel__pk",
-        "slug": "channel_listing__channel__slug",
-        "currency_code": "channel_listing__channel__currency_code",
-        "published": "channel_listing__is_published",
-        "publication_date": "channel_listing__publication_date",
-    }
-
-    WAREHOUSE_FIELDS = {
-        "slug": "variants__stocks__warehouse__slug",
-        "quantity": "variants__stocks__quantity",
-        "warehouse_pk": "variants__stocks__warehouse__id",
-    }
-
-    VARIANT_ATTRIBUTE_FIELDS = {
-        "value": "variants__attributes__values__slug",
-        "slug": "variants__attributes__assignment__attribute__slug",
-        "attribute_pk": "variants__attributes__assignment__attribute__pk",
-    }
-
-    VARIANT_CHANNEL_LISTING_FIELDS = {
-        "channel_pk": "variants__channel_listing__channel__pk",
-        "slug": "variants__channel_listing__channel__slug",
-        "price_amount": "variants__channel_listing__price_amount",
-        "currency": "variants__channel_listing__currency",
-    }
-
-
-def get_export_fields_and_headers_info(
-    export_info: Dict[str, list]
-) -> Tuple[List[str], List[str], List[str]]:
-    """Get export fields, all headers and headers mapping.
-
-    Based on export_info returns exported fields, fields to headers mapping and
-    all headers.
-    Headers contains product, variant, attribute and warehouse headers.
-    """
-    export_fields, file_headers = get_product_export_fields_and_headers(export_info)
-    attributes_headers = get_attributes_headers(export_info)
-    warehouses_headers = get_warehouses_headers(export_info)
-    channels_headers = get_channels_headers(export_info)
-
-    data_headers = (
-        export_fields + attributes_headers + warehouses_headers + channels_headers
-    )
-    file_headers += attributes_headers + warehouses_headers + channels_headers
-    return export_fields, file_headers, data_headers
-
-
-def get_product_export_fields_and_headers(
-    export_info: Dict[str, list]
-) -> Tuple[List[str], List[str]]:
-    """Get export fields from export info and prepare headers mapping.
-
-    Based on given fields headers from export info, export fields set and
-    headers mapping is prepared.
-    """
-    export_fields = ["id"]
-    file_headers = ["id"]
-
-    fields = export_info.get("fields")
-    if not fields:
-        return export_fields, file_headers
-
-    fields_mapping = dict(
-        ChainMap(*reversed(ProductExportFields.HEADERS_TO_FIELDS_MAPPING.values()))  # type: ignore
-    )
-
-    for field in fields:
-        lookup_field = fields_mapping[field]
-        export_fields.append(lookup_field)
-        file_headers.append(field)
-        # if price is exported, currency is needed too
-        if field == "cost price":
-            lookup_field = fields_mapping["variant currency"]
-            export_fields.append(lookup_field)
-            file_headers.append("variant currency")
-
-    return export_fields, file_headers
-
-
-def get_attributes_headers(export_info: Dict[str, list]) -> List[str]:
-    """Get headers for exported attributes.
-
-    Headers are build from slug and contains information if it's a product or variant
-    attribute. Respectively for product: "slug-value (product attribute)"
-    and for variant: "slug-value (variant attribute)".
-    """
-
-    attribute_ids = export_info.get("attributes")
-    if not attribute_ids:
-        return []
-
-    attributes = Attribute.objects.filter(pk__in=attribute_ids).order_by("slug")
-
-    products_headers = (
-        attributes.filter(product_types__isnull=False)
-        .annotate(header=Concat("slug", V(" (product attribute)")))
-        .values_list("header", flat=True)
-    )
-
-    variant_headers = (
-        attributes.filter(product_variant_types__isnull=False)
-        .annotate(header=Concat("slug", V(" (variant attribute)")))
-        .values_list("header", flat=True)
-    )
-
-    return list(products_headers) + list(variant_headers)
-
-
-def get_warehouses_headers(export_info: Dict[str, list]) -> List[str]:
-    """Get headers for exported warehouses.
-
-    Headers are build from slug. Example: "slug-value (warehouse quantity)"
-    """
-    warehouse_ids = export_info.get("warehouses")
-    if not warehouse_ids:
-        return []
-
-    warehouses_headers = (
-        Warehouse.objects.filter(pk__in=warehouse_ids)
-        .order_by("slug")
-        .annotate(header=Concat("slug", V(" (warehouse quantity)")))
-        .values_list("header", flat=True)
-    )
-
-    return list(warehouses_headers)
-
-
-def get_channels_headers(export_info: Dict[str, list]) -> List[str]:
-    """Get headers for exported channels.
-
-    Headers are build from slug and exported field.
-
-    Example:
-    - currency code data header: "slug-value (channel currency code)"
-    - published data header: "slug-value (channel visible)"
-    - publication date data header: "slug-value (channel publication date)"
-
-    """
-    channel_ids = export_info.get("channels")
-    if not channel_ids:
-        return []
-
-    channels_slugs = (
-        Channel.objects.filter(pk__in=channel_ids)
-        .order_by("slug")
-        .values_list("slug", flat=True)
-    )
-
-    fields = [
-        *ProductExportFields.PRODUCT_CHANNEL_LISTING_FIELDS.keys(),
-        *ProductExportFields.VARIANT_CHANNEL_LISTING_FIELDS.keys(),
-    ]
-    channels_headers = []
-    for slug in channels_slugs:
-        channels_headers.extend(
-            [
-                f"{slug} (channel {field.replace('_', ' ')})"
-                for field in fields
-                if field not in ["slug", "channel_pk"]
-            ]
-        )
-
-    return list(channels_headers)
 
 
 def get_products_data(
